+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `ip_http_secure_active_session_modules`, `ip_http_max_connections` and `ip_http_active_session_modules` attributes to `iosxe_system` resource and data source
 - Fix idempotency issue with `key` attribute of `iosxe_tacacs_server` resource
 - Add `encryption` attribute to `iosxe_tacacs_server` resource and data source
+- Add `trusted_keys` attribute to `iosxe_ntp` resource and data source
 
 ## 0.5.8